+++ packages/core/bootstrap/src/lib/external-adapter/errors.ts
@@ -10,6 +10,7 @@ export class AdapterError extends Error {
   url?: string
   errorResponse: any
   feedID?: string
+  providerStatusCode?: number
 
   constructor({
     jobRunID = '1',
@@ -21,6 +22,7 @@ export class AdapterError extends Error {
     url,
     errorResponse,
     feedID,
+    providerStatusCode,
   }: Partial<AdapterError>) {
     super(message)
 
@@ -37,6 +39,7 @@ export class AdapterError extends Error {
       this.feedID = feedID
     }
     this.errorResponse = errorResponse
+    this.providerStatusCode = providerStatusCode
   }
 
   toJSONResponse(): AdapterErrorResponse {
@@ -53,6 +56,7 @@ export class AdapterError extends Error {
       jobRunID: this.jobRunID,
       status: this.status,
       statusCode: this.statusCode,
+      providerStatusCode: this.providerStatusCode,
       error: showDebugInfo ? errorFull : errorBasic,
     }
   }