+++ packages/sources/coincodex/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as getcoin from './getcoin'