+++ packages/sources/terra-view-function/src/endpoint/view.ts
@@ -6,11 +6,26 @@ import { Config, ChainId, SUPPORTED_CHAIN_IDS } from '../config'
 export const supportedEndpoints = ['view']
 
 export const inputParameters: InputParameters = {
-  address: ['address', 'contract'],
-  query: true,
-  params: false,
-  chainId: false,
-  resultPath: false,
+  address: {
+    aliases: ['contract'],
+    required: true,
+    description: 'The address to query',
+    type: 'string',
+  },
+  query: {
+    required: true,
+    description: 'The query object',
+  },
+  params: {
+    required: false,
+    description: 'Optional params object to include in the query',
+  },
+  chainId: {
+    required: false,
+    description:
+      'Which chain ID to connect to. Default is `DEFAULT_CHAIN_ID` environment variable (`columbus-5`, `bombay-12`, `localterra`, etc.)',
+    type: 'string',
+  },
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {